better attention model
Now, a conventional RNN is used for attention (thus helping to ensure that the output at a given timestep is highly dependent on that timestep). The attention mechanism is also given a higher "temperature", allowing the softmax to show more initial variation. It is not clear how much the latter change affects the result.

Experimental results:
- After 15000 iterations with batch size 4 and step 0.001, the average batch cost was 6.6
- After an additional 500 iterations with step 0.0001, it was 2.32.
- The attention moves while looking at a sample, rather than getting stuck at one place.